crossbuild: build the libc db code during bootstrap on MacOS and Linux
The Linux db functions create databases that are incompatible with what
libc handles so the databases created during the install phase will not be
readable unless we use the same db code to create them.
To ensure that the bootstrap tools always link against this library instead
of the system one I define dbopen() to __freebsd_dbopen() before including
db.h so that we get linker errors instead of accidentally linking against
the host dbopen().