Stop using unifdef to generate bsdxml.h
Summary:
The current invocation of unifdef causes the build to fail when using a shell with -o pipefail on by default since unifdef will return a non-zero exit status if it changes something. The only thing this call to unifdef does is remove 5 lines that will be ignored by the compiler anyway. Furthermore, it is the only make rule in the source tree that requires unifdef. Removing this call also makes it slightly easier to build without inhering $PATH (D16815) since we don't need unifdef anymore.

I also noticed that the sed call to replace the include guard has been broken for over 10 years since the import of expat 2.0.1 changed it from `XmlParse_INCLUDED` to `Expat_INCLUDED`. I could also fix this but since it's been broken for so long and no one noticed, it's probably not necessary.

Reviewers: emaste, jhb, imp, bdrewery

Differential Revision: https://reviews.freebsd.org/D14317