Don't run cc --version during cleandir/obj stages
Summary:
This will no work when there is no cc in $PATH (which is the case before the
cross-tools stage once we no longer inherit $PATH in $WMAKE).
The variables set by bsd.compiler.mk/bsd.linker.mk are not needed in these
stages so this avoids a little bit of makefile parsing.

Test Plan: buildworld works with my patches to remove $PATH from $TMPPATH

Reviewers: jhb, bdrewery

Subscribers: emaste

Differential Revision: https://reviews.freebsd.org/D16814