Fix building when using a shell with -o pipefail for bmake
Summary:
When running ldd some of the files (e.g. make) will be statically linked
so ldd will return a non-zero exit status. With -o pipefail this will fail
the whole rule. The source directory for find in delete-old-files might
also not exist so add a || true to make it pass

Reviewers: brooks, emaste, jhb

Differential Revision: https://reviews.freebsd.org/D16864