Add a BUILD_WITH_OPIPEFAIL option to use `bash -o pipefail` as the bmake shell
Summary:
This is useful to find failed commands if the output is piped. By default
shells will only return the exit code of the last command in the pipeline.
The -o pipefail option causes the pipelline exit code to be nonzero if any
of the intermediate commands return non-zero.
This has been extremely useful for me when debugging build failures on
non-FreeBSD hosts due to missing/incompatible commands.

This currently creates a sh binary that will execute `bash -o pipefail` but
once we have -o pipefail support in /bin/sh we could instead bootstrap a
version of /bin/sh that defaults to pipefail.

Reviewers: jhb, brooks, emaste, imp, bdrewery

Differential Revision: https://reviews.freebsd.org/D16868