Don't delete the source file when installing links on case-insensitive file systems
Summary:
If the name of the link is the same except for case, use install -S to
avoid removing the destination file (which will be the source on a
case-insensitive file system) before creating the link.
I've also modified install.sh to skip the ln call if the file names are
the same in lower case and test -ef (same file) returns true.

Without this change installworld fails on MacOS in libcrypto (due to hmac.3
-> HMAC.3) and sendmail (mail -> Mail)

Test Plan: installworld works on MacOS (with -DNO_ROOT)

Reviewers: emaste, jhb, bdrewery, imp

Differential Revision: https://reviews.freebsd.org/D16762