Add bootstrap files to m4 to allow building without a host m4
This uses the same approach as usr.bin/lex. This is required when building
without using the host $PATH since there is a cyclic dependency between
lex and m4. m4 is needed to build lex and m4 uses lex to generate files.
It is already possible to bootstrap lex, but lex will call m4 when generating
files so it can't be used to bootstrap m4.
For now use `make bootstrap` for both, but it should be enough to bootstrap
m4 and then use that to build lex