Check for /usr/bin/tsort /usr/bin/lorder when bootstrapping on non-FreeBSD
This fixes empty .a files being create if one of of the tools is missing.
lorder is usually not installed by default on Linux systems and using the
tools is not be required if we link with lld, so it might make sense to
require lld as a host linker when building on Linux.