Build the elftoolchain libs as part of cross-tools instead of bootstrap-tools
To build them we need m4 installed which is only built in the
bootstrap-tools phase if we set BOOTSTRAP_ALL_TOOLS. I don't see any
reason why libelf needs to be built as part of legacy instead of
bootstrap-tools so move it to bootstrap-tools instead.
This also enables greater parallelism since the legacy stage is serial.