Always rebuild tables.h for libsysdecode
The file depends on basically all system headers so we should just
rebuild it every time. However, it is rarely going to change so we can use
the same approach as for ioctl.c and only update tables.h if there is
actually a change. This avoids recompiling and relinking libsysdecode if
nothing changed.