build: Makefile.boot: Set LORDER only if not already defined
lorder is a BSD-specific script which is part of the FreeBSD source
code. For linux, a separate package is needed or the script can be
copied from source to /usr/bin/lorder. Both approaches require root
access.

With this patch, the user can set the location for lorder (e.g the
FreeBSD source code) by using the environment variable LORDER.