make.py: Rework bootstrap_* to be self-contained
bootstrap_bmake() and bootstrap_libbsd() have been rewritten to be
self-contained and to avoid sharing global state via global variables.
Several checks related to the existence of the packages have also been
moved to the corresponding boostrap_* functions.

Part of the patch is a subtle fix for MAKEOBJDIRPREFIX. Right before
starting the build process, the current directory is switched to the
sources root directory. This breaks MAKEOBJDIRPREFIX if it was
specified as a relative path instead of an absolute path. Use
MAKEOBJDIRPREFIX as an absolute path instead.