Allow linking the kernel with upstream LLD
Upstream LLD does not support the -z ifunc-noplt flag and since version
7.0.0 it will give an error when passed an unsupported -z option instead
of ignoring it.
This commit adds a new LINKER_FEATURE ifunc-noplt based on the value of
LINKER_FREEBSD_VERSION and only adds the flag if we know it is supported.
In case it isn't support a .warning was added to inform users that the
kernel will not be quite as fast as it could be if it was linked with the
ifunc-noplt option.