Skip LIB32 when building with upstream lld 7.0
When linking with upstream LLD 7.0 we get the following error for lib32:
.../tmp/usr/lib32/crt1.o: index past the end of the symbol table

This was fixed in https://reviews.llvm.org/rL345002 (will be in 7.1).

I initially tried to do this in src.opts.mk by adding LIB32 to BROKEN
but it seems that it uses the host linker rather than the target one at the
point when the LIB32 decision is made. To work around this I added the
logic to Makefile.inc1 instead.