Stop using ld -r + objcopy to create i386 crt1.o
I am currently trying to build using upstream LLVM 7.0 and when linking any
i386 binary with upstream LLD 7.0 we get the following error:
.../tmp/usr/lib32/crt1.o: index past the end of the symbol table

This was fixed in https://reviews.llvm.org/rL345002 (will be in 7.0.1).
However, that version of LLVM has not been released yet so I decided to fix
the problem by no longer using ld -r. To achieve this we now include the
crt1_s.S file as inline assembly inside the C file.

While this is slightly less readable, it has the advantage of no longer
needing ld -r and also since _start is now defined in the same file we
no longer need to run ${OBJCOPY} --localize-symbol _start1.