Combined patch to allow cross-building on Linux/macOS hosts
Usage: instead of running bmake in the source directory run
`./tools/build/make.py` instead. This will bootstrap bmake and
perform some sanity checks.
It may also be possible to use a distribution-provided bmake binary, but
this has not been tested.

Many of these changes have upstream reviews, so it may result in merge
conflicts when updating. However, it is probably easier to resolve
conflicts in a single commit rather than a range of commits.