Add makefs to the list of bootstrap tools
Having this as part of the bootstrap tools is useful to build disk images
to boot in QEMU (especially when building on a Linux/macOS host where
mkimg is not available). We have been bootstrapping makefs+mkimg in
CheriBSD for a long time (using LOCAL_XTOOL_DIRS) but I believe this
is also useful upstream.

Reviewers: emaste, bdrewery, brooks, imp