truss: fix handling of 64-bit arguments/return values for compat32
Summary:
Deciding whether to combine two values to a 64-bit one should be based on
the process ABI, and not dependent on whether truss is compiled for an
LP64 ABI. This is a follow-up cleanup for D27625. I found this while
looking for uses of the __LP64__ macro (since using this is wrong for
CHERI systems).

Test Plan: still works. Since I tested on AMD64 and all syscalls in the trace have their quad argument last there was no difference in the output. Should fix output for compat32 on MIPS64 though.

Reviewers: jhb

Subscribers: imp

Differential Revision: https://reviews.freebsd.org/D27637