libc: Add an ASAN-compatible strlen implementation
This change was taken from CheriBSD (commit xxxxx) since it is also
required for CHERI. Without this change we get ASAN errors very early on
(during jemalloc startup) and can't use an instrumented libc.

The version that reads multiple bytes at a time is still the default, but
we pick the safe version when compiling with -fsanitize=address.